consolidate virtual pin logic
